requirements.txt
python-mpd2==3.0.4
